+++ backend/index.js
@@ -9,15 +9,14 @@ app.use(express.json());
 const userEndpoints = require("./routes/userRoutes.js");
 const movieEndpoints = require("./routes/movieRoutes.js");
 const revEndpoints = require("./routes/reviewRoutes.js");
-const { login, signup } = require("./routes/auth.js");
+const authEndpoints = require("./routes/authRoutes.js");
 
 app.use(logger);
 
 app.use("/movies", movieEndpoints);
 app.use("/users", userEndpoints);
 app.use("/reviews", revEndpoints);
-app.post("/login", login);
-app.post("/signup", signup);
+app.use("/auth", authEndpoints);
 
 function logger(req, res, next) {
   console.log(`${req.method} ${req.path}`);
